# swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/registro":{"post":{"operationId":"RegistroController_post","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistroDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"RegistroController_getAll","parameters":[{"name":"query","required":false,"in":"query","description":"query - Filter. e.g. col1:v1,col2:v2 …","schema":{"type":"string"}},{"name":"fields","required":false,"in":"query","description":"fields - Fields returned. e.g. col1,col2 …","schema":{"type":"string"}},{"name":"sortby","required":false,"in":"query","description":"sortby - Sorted-by fields. e.g. col1,col2 …","schema":{"type":"string"}},{"name":"order","required":false,"in":"query","description":"order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","description":"limit - Limit the size of result set. Must be an integer","schema":{"type":"string"}},{"name":"offset","required":false,"in":"query","description":"offset - Start position of result set. Must be an integer","schema":{"type":"string"}},{"name":"populate","required":false,"in":"query","description":"populate - show subqueries. Must be a boolean","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/registro/{id}":{"get":{"operationId":"RegistroController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"put":{"operationId":"RegistroController_put","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistroDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"RegistroController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/nota":{"post":{"operationId":"NotaController_post","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotaDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"NotaController_getAll","parameters":[{"name":"query","required":false,"in":"query","description":"query - Filter. e.g. col1:v1,col2:v2 …","schema":{"type":"string"}},{"name":"fields","required":false,"in":"query","description":"fields - Fields returned. e.g. col1,col2 …","schema":{"type":"string"}},{"name":"sortby","required":false,"in":"query","description":"sortby - Sorted-by fields. e.g. col1,col2 …","schema":{"type":"string"}},{"name":"order","required":false,"in":"query","description":"order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","description":"limit - Limit the size of result set. Must be an integer","schema":{"type":"string"}},{"name":"offset","required":false,"in":"query","description":"offset - Start position of result set. Must be an integer","schema":{"type":"string"}},{"name":"populate","required":false,"in":"query","description":"populate - show subqueries. Must be a boolean","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/nota/{id}":{"get":{"operationId":"NotaController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"put":{"operationId":"NotaController_put","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotaDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"NotaController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"registro_notas_crud","description":"API CRUD para el registro de notas","version":"1.0","contact":{}},"tags":[{"name":"registro_nota","description":""}],"servers":[],"components":{"schemas":{"RegistroDto":{"type":"object","properties":{"nombre":{"type":"string"},"descripcion":{"type":"string"},"codigo_abreviacion":{"type":"string"},"codigo":{"type":"string"},"periodo_id":{"type":"string"},"nivel_id":{"type":"string"},"espacio_academico_id":{"type":"string"},"estado_registro_id":{"type":"string"},"estructura_nota":{"type":"object"},"activo":{"type":"boolean"},"fecha_creacion":{"format":"date-time","type":"string"},"fecha_modificacion":{"format":"date-time","type":"string"}},"required":["nombre","descripcion","codigo_abreviacion","codigo","periodo_id","nivel_id","espacio_academico_id","estado_registro_id","estructura_nota","activo","fecha_creacion","fecha_modificacion"]},"NotaDto":{"type":"object","properties":{"nombre":{"type":"string"},"descripcion":{"type":"string"},"estudiante_id":{"type":"string"},"registro_id":{"type":"string"},"valor_nota":{"type":"object"},"nota_definitiva":{"type":"number"},"aprobado":{"type":"boolean"},"homologado":{"type":"boolean"},"fallas":{"type":"number"},"observacion_nota_id":{"type":"string"},"activo":{"type":"boolean"},"fecha_creacion":{"format":"date-time","type":"string"},"fecha_modificacion":{"format":"date-time","type":"string"}},"required":["nombre","descripcion","estudiante_id","registro_id","valor_nota","nota_definitiva","aprobado","homologado","fallas","observacion_nota_id","activo","fecha_creacion","fecha_modificacion"]}}}}